requirements.txt
asgiref==3.11.0
Django==5.2.8
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
sqlparse==0.5.3
